added new features to enable plot and export to excel

+++ PseudoSkinApplication/RunSensititvity/RunSensitivityDto.cs
@@ -2,7 +2,7 @@
 {
     public class RunSensitivityDto
     {
-        public int SerialNo { get; set; }
+        public double SerialNo { get; set; }
         public double ParameterValue { get; set; }
         public double Pseudoskin { get; set; }
     }
